requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.7.0
bokeh==3.3.1
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
distro==1.9.0
easygui==0.98.3
Flask==3.0.3
Flask-Session==0.8.0
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.6.1
lxml==5.3.0
MarkupSafe==2.1.3
msgspec==0.18.6
multidict==6.1.0
oauthlib==3.2.2
openai==1.51.2
ordered-set==4.1.0
packaging==23.2
pandas==2.1.2
pefile==2023.2.7
Pillow==10.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-docx==1.1.2
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
style==1.1.0
tornado==6.3.3
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2023.3
update==0.0.1
urllib3==2.2.2
Werkzeug==3.0.2
win_unicode_console==0.5
xyzservices==2023.10.1
yarl==1.14.0
zstandard==0.22.0
